Refactor createLeveredPosition to a hook
