+++ src/hooks/useLeveredPositions.tsx
@@ -1,0 +1,39 @@
+import { useCallback } from 'react';
+import { BigNumber } from 'ethers';
+import { sendTransaction } from '../actions';
+import { buildBuyCollateralAndIncreaseLeverArgs } from '../actions';
+import { useAddRecentTransaction } from '@rainbow-me/rainbowkit';
+import { chain as chains, useAccount, useNetwork } from 'wagmi';
+import useStore from '../state/stores/globalStore';
+import useSoftReset from './useSoftReset';
+import { useUserData } from '../state/queries/useUserData';
+
+export const useCreateLeveredPosition = () => {
+
+  const addRecentTransaction = useAddRecentTransaction();
+  const { chain } = useNetwork();
+  const { address } = useAccount();
+  const softReset = useSoftReset();
+
+  const fiat = useStore(state => state.fiat);
+  const user = useStore((state) => state.user);
+  const modifyPositionData = useStore((state) => state.modifyPositionData);
+
+  const { data: userData } = useUserData(fiat, chain?.id ?? chains.mainnet.id, address ?? '');
+  const { proxies } = userData as any;
+  
+  const createLeveredPosition = useCallback(async (
+    upFrontUnderlier: BigNumber, addDebt: BigNumber, minUnderlierToBuy: BigNumber, minTokenToBuy: BigNumber
+  ) => {
+    const args = await buildBuyCollateralAndIncreaseLeverArgs(
+      fiat, user, proxies, modifyPositionData.collateralType, upFrontUnderlier, addDebt, minUnderlierToBuy, minTokenToBuy
+    );
+    const response = await sendTransaction(
+      fiat, true, proxies[0], 'createLeveredPosition', args.contract, args.methodName, ...args.methodArgs
+    );
+    addRecentTransaction({ hash: response.transactionHash, description: 'Create levered position' });
+    softReset();
+  }, [addRecentTransaction, fiat, modifyPositionData.collateralType, proxies, softReset, user]);
+
+  return createLeveredPosition;
+}
